move unaligned pointers test from `lists` to `invalid`
