+++ tests/runtime/invalid/exports.wit
@@ -5,3 +5,5 @@ invalid-s16: func()
 invalid-char: func()
 invalid-bool: func()
 invalid-enum: func()
+
+test-unaligned: func()
