+++ tests/runtime/invalid/host.py
@@ -27,6 +27,21 @@ class MyImports(Imports):
     def roundtrip_enum(self, x: i.E) -> i.E:
         raise Exception('unreachable')
 
+    def unaligned_roundtrip1(self, a: List[int], b: List[int], c: List[int], d: List[i.Flag32], e: List[i.Flag64]) -> None:
+        assert(a == [1])
+        assert(b == [2])
+        assert(c == [3])
+        assert(d == [i.Flag32.B8])
+        assert(e == [i.Flag64.B9])
+
+    def unaligned_roundtrip2(self, a: List[i.UnalignedRecord], b: List[float], c: List[float], d: List[str], e: List[bytes]) -> None:
+          assert(a == [i.UnalignedRecord(a=10, b=11)])
+          assert(b == [100.0])
+          assert(c == [101.0])
+          assert(d == ['foo'])
+          assert(e == [b'\x66'])
+
+
 def run(wasm_file: str) -> None:
     store = wasmtime.Store()
     module = wasmtime.Module.from_file(store.engine, wasm_file)
@@ -64,6 +79,7 @@ def run(wasm_file: str) -> None:
     assert_throws(lambda: wasm.invalid_s16(store), 'must be between')
     assert_throws(lambda: wasm.invalid_char(store), 'not a valid char')
     assert_throws(lambda: wasm.invalid_enum(store), 'not a valid E')
+    assert_throws(lambda: wasm.test_unaligned(store), 'is not aligned');
 
 if __name__ == '__main__':
     run(sys.argv[1])
