+++ tests/runtime/lists/host.rs
@@ -95,38 +95,6 @@ impl Imports for MyImports {
         assert_eq!(s, [f64::MIN, f64::MAX, f64::NEG_INFINITY, f64::INFINITY]);
         (u, s)
     }
-
-    fn unaligned_roundtrip1(
-        &mut self,
-        u16s: Vec<u16>,
-        u32s: Vec<u32>,
-        u64s: Vec<u64>,
-        flag32s: Vec<Flag32>,
-        flag64s: Vec<Flag64>,
-    ) {
-        assert_eq!(u16s, [1]);
-        assert_eq!(u32s, [2]);
-        assert_eq!(u64s, [3]);
-        assert_eq!(flag32s, [Flag32::B8]);
-        assert_eq!(flag64s, [Flag64::B9]);
-    }
-
-    fn unaligned_roundtrip2(
-        &mut self,
-        records: Vec<UnalignedRecord>,
-        f32s: Vec<f32>,
-        f64s: Vec<f64>,
-        strings: Vec<String>,
-        lists: Vec<Vec<u8>>,
-    ) {
-        assert_eq!(records.len(), 1);
-        assert_eq!(records[0].a, 10);
-        assert_eq!(records[0].b, 11);
-        assert_eq!(f32s, [100.0]);
-        assert_eq!(f64s, [101.0]);
-        assert_eq!(strings, ["foo"]);
-        assert_eq!(lists, [&[102][..]]);
-    }
 }
 
 wit_bindgen_host_wasmtime_rust::import!("../../tests/runtime/lists/exports.wit");
